Add file upload support in `EditDocumentForm`
Refactor `SingleFileUploader` to `FileUploader` with upload functionality. Integrated new upload handler in `EditDocumentForm` to handle document uploads, including toast notifications upon success or failure. Simplified async call in `postFormData` to return fetch promise.
